fix(prompts): scope Gemini COMMENTED review to latest only; add reviews fetch to check_existing_pr
- review_prompt step 2b: select only the MOST RECENT Gemini COMMENTED
  review (sort by submitted_at, take last); older reviews are superseded
  and must be ignored — prevents stale body-only feedback from blocking
  the review loop after a re-review
- check_existing_pr: add steps 2/2b/3/3b mirroring review_prompt so
  direct PR tasks (e.g. @harness review) also fetch /reviews and detect
  Gemini body-only COMMENTED feedback via trigger phrases
- Split prompts/review.rs into review/mod.rs + review/tests.rs to stay
  within the 800-line file limit
- Add tests: test_review_prompt_gemini_body_only_instructions,
  test_review_prompt_gemini_stale_filter,
  test_check_existing_pr_fetches_reviews_and_body_feedback
